refactor: 요금 계산 기능 리팩토링
