+++ src/main/java/nextstep/subway/applicaion/dto/PathResponse.java
@@ -2,6 +2,7 @@ package nextstep.subway.applicaion.dto;
 
 import lombok.Getter;
 import nextstep.subway.domain.Path;
+import nextstep.subway.domain.Price;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -12,12 +13,6 @@ public class PathResponse {
     private int distance;
     private int duration;
     private int price;
-
-    public PathResponse(List<StationResponse> stations, int distance, int duration) {
-        this.stations = stations;
-        this.distance = distance;
-        this.duration = duration;
-    }
 
     public PathResponse(List<StationResponse> stations, int distance, int duration, int price) {
         this.stations = stations;
@@ -32,7 +27,7 @@ public class PathResponse {
                 .collect(Collectors.toList());
         int distance = path.extractDistance();
         int duration = path.extractDuration();
-        int price = path.calculatePrice();
-        return new PathResponse(stations, distance, duration);
+        int price = Price.calculate(distance).getPrice();
+        return new PathResponse(stations, distance, duration, price);
     }
 }
